requirements.txt
APScheduler==3.10.4
asn1crypto==0.24.0
attrs==23.1.0
Automat==0.6.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
click==6.7
cloud-init==23.1.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
exceptiongroup==1.1.3
h11==0.14.0
httplib2==0.9.2
hyperlink==17.3.1
idna==3.4
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
outcome==1.3.0.post0
packaging==23.2
PAM==0.4.2
pexpect==4.2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
PySocks==1.7.1
python-apt==1.6.6
python-debian==0.1.32
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.25
PyYAML==3.12
requests==2.31.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
selenium==4.14.0
service-identity==16.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.4
ssh-import-id==5.7
systemd-python==234
trio==0.22.2
trio-websocket==0.11.1
Twisted==17.9.0
tzlocal==5.2
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
zope.interface==4.3.2
